+++ lib/blocs/push_notification/push_notification_bloc.dart
@@ -7,7 +7,6 @@ import 'package:device_info_plus/device_info_plus.dart';
 import 'package:equatable/equatable.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 import 'package:get_it/get_it.dart';
 import 'package:supa_architecture/repositories/utils_notification_repository.dart';
 import 'package:supa_architecture/supa_architecture.dart';
@@ -24,13 +23,8 @@ class PushNotificationBloc
 
   StreamSubscription? _notificationOpenSubscription;
 
-  /// Show local notification
-  final bool showLocalNotification;
-
   /// Push notification BLoC for handling push notifications.
-  PushNotificationBloc({
-    this.showLocalNotification = false,
-  }) : super(const PushNotificationInitial()) {
+  PushNotificationBloc() : super(const PushNotificationInitial()) {
     on<DidReceivedNotificationEvent>(_onDidNotificationReceived);
     on<DidUserOpenedNotificationEvent>(_onDidUserOpenedNotification);
     on<DidResetNotificationEvent>(_onDidResetNotification);
@@ -88,10 +82,6 @@ class PushNotificationBloc
   /// Firebase messaging instance
   FirebaseMessaging get _firebaseMessaging => FirebaseMessaging.instance;
 
-  /// Flutter local notifications instance
-  final FlutterLocalNotificationsPlugin _flutterLocalNotificationsPlugin =
-      FlutterLocalNotificationsPlugin();
-
   /// Initialize push notifications (with default channel for Android).
   /// - Uses app id as the channel id.
   Future<void> initializeNotifications({
@@ -103,12 +93,6 @@ class PushNotificationBloc
 
     // Initialize Firebase Messaging
     await _initializeFirebaseMessaging();
-
-    // Initialize Local Notifications
-    await _initializeLocalNotifications(
-      channelId: appId,
-      channelName: channelName,
-    );
 
     /// Register device token for notifications.
     await registerDeviceToken();
@@ -177,11 +161,6 @@ class PushNotificationBloc
   /// Handle foreground notifications by showing a local notification
   /// and passing it to the specified message handler.
   void _handleForegroundNotification(RemoteMessage message) async {
-    if (showLocalNotification) {
-      // Show local notification to indicate incoming message.
-      await _showLocalNotification(message);
-    }
-
     add(
       DidReceivedNotificationEvent(
         title: message.notification?.title ?? '',
@@ -197,88 +176,6 @@ class PushNotificationBloc
     await _firebaseMessaging.requestPermission();
     String? token = await _firebaseMessaging.getToken();
     debugPrint("FIREBASE_MESSAGING_TOKEN: $token");
-  }
-
-  /// Initialize local notifications (only for Android).
-  Future<void> _initializeLocalNotifications({
-    required String channelId,
-    String? channelName,
-  }) async {
-    final InitializationSettings initializationSettings =
-        InitializationSettings(
-      android: const AndroidInitializationSettings(
-        '@mipmap/ic_launcher',
-      ),
-      iOS: DarwinInitializationSettings(
-        requestAlertPermission: true,
-        requestBadgePermission: true,
-        requestSoundPermission: true,
-        onDidReceiveLocalNotification: _onDidReceiveLocalNotification,
-      ),
-    );
-
-    await _flutterLocalNotificationsPlugin.initialize(
-      initializationSettings,
-      onDidReceiveNotificationResponse: _onDidReceivedNotificationResponse,
-    );
-
-    // Set the default notification channel on Android.
-    if (Platform.isAndroid) {
-      final AndroidNotificationChannel channel = AndroidNotificationChannel(
-        channelId,
-        channelName ?? channelId,
-        importance: Importance.high,
-      );
-
-      await _flutterLocalNotificationsPlugin
-          .resolvePlatformSpecificImplementation<
-              AndroidFlutterLocalNotificationsPlugin>()
-          ?.createNotificationChannel(channel);
-    }
-  }
-
-  void _onDidReceivedNotificationResponse(
-    NotificationResponse notificationResponse,
-  ) {
-    debugPrint(
-      "Received notification response: ${notificationResponse.id}, ${notificationResponse.payload}",
-    );
-  }
-
-  void _onDidReceiveLocalNotification(
-    int id,
-    String? title,
-    String? body,
-    String? payload,
-  ) {
-    debugPrint(
-      "Received notification: id: $id, title: $title, body: $body, payload: $payload",
-    );
-  }
-
-  /// Show a local notification on the device (use for foreground notifications).
-  Future<void> _showLocalNotification(
-    RemoteMessage message, {
-    String channelId = 'default_channel',
-    String? channelName,
-  }) async {
-    final AndroidNotificationDetails androidPlatformChannelSpecifics =
-        AndroidNotificationDetails(
-      channelId,
-      channelName ?? channelId,
-      importance: Importance.max,
-      priority: Priority.high,
-    );
-
-    final NotificationDetails platformChannelSpecifics =
-        NotificationDetails(android: androidPlatformChannelSpecifics);
-
-    await _flutterLocalNotificationsPlugin.show(
-      message.notification?.hashCode ?? 0,
-      message.notification?.title,
-      message.notification?.body,
-      platformChannelSpecifics,
-    );
   }
 
   /// Private method to check if the device is Android 13 or higher.