+++ lib/models/attachment.dart
@@ -1,6 +1,6 @@
-import 'package:supa_architecture/data/app_user.dart';
-import 'package:supa_architecture/data/file.dart';
 import 'package:supa_architecture/json/json.dart';
+import 'package:supa_architecture/models/app_user.dart';
+import 'package:supa_architecture/models/file.dart';
 
 /// An abstract class representing an attachment model.
 ///